HoverMindAI/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
openai==1.3.7
SpeechRecognition==3.10.0
python-dotenv==1.0.0 